fix: address review feedback on VoiceStatusPanel
Exclude protected users from selection-mode bulk actions, reseed join
times and voice-state baselines for existing channel members on
select, use Constants.Endpoints for guild member requests instead of
raw route strings, and split settings/icons/pure helpers into their
own files.
